fix(build-order): preserve raw supply and clamp display to 200
Storage was u8, so food_used/food_made > 255 wrapped mod 256 (firebat
golden had silent 271→15 and 256→0); 200–255 leaked past the in-game
cap (e.g. 217/202). Widen to u16, drop the lossy cast, and clamp
supply_made.min(200) only at render/export — matching the Timeline
side panel and SC2 HUD while keeping the struct faithful to the tracker.

+++ src/build_order/types.rs
@@ -37,9 +37,11 @@ impl EntryOutcome {
 #[derive(Clone)]
 pub struct BuildOrderEntry {
     /// Supply usado no instante de início.
-    pub supply: u8,
+    pub supply: u16,
     /// Capacidade total de supply no instante de início (food_made).
-    pub supply_made: u8,
+    /// Pode exceder 200: late-game Zerg/Protoss costuma ter overlords/pylons
+    /// além do cap visual. Quem renderiza decide se aplica `min(200)`.
+    pub supply_made: u16,
     /// Instante de início da ação (start time).
     pub game_loop: u32,
     /// Instante de conclusão da ação. Significado depende do `outcome`:
